requirements.txt
pyinstaller
pyinstaller-hooks-contrib
pywebview
cefpython3